taint/test/mochitest/test_message_event.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Check Location Sources</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      SimpleTest.waitForExplicitFinish();
      function startTest() {
          const myWorker = new SharedWorker("mock_worker.js");

          myWorker.port.onmessage = function (event) {
              check_tainted(event.data);
              SimpleTest.info("hello there");
              SimpleTest.finish();
          }

/*
          myWorker.onconnect = (event) => {
              const port = event.ports[0]; 
              myWorker.port.postMessage("test");
          }*/


          myWorker.port.postMessage("test");
      }

    </script>
  </head>
  <body onload="startTest()">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</html>
